adding some things were missing

+++ src/main/java/com/truper/test/dto/ProductosDto.java
@@ -1,10 +1,12 @@
 package com.truper.test.dto;
 
+import java.math.BigDecimal;
+
 public class ProductosDto {
 	
 	private String codigo;
 	private String descripcion;
-	private int precio;
+	private BigDecimal precio;
 	
 	
 	public String getCodigo() {
@@ -19,10 +21,10 @@ public class ProductosDto {
 	public void setDescripcion(String descripcion) {
 		this.descripcion = descripcion;
 	}
-	public int getPrecio() {
+	public BigDecimal getPrecio() {
 		return precio;
 	}
-	public void setPrecio(int precio) {
+	public void setPrecio(BigDecimal precio) {
 		this.precio = precio;
 	}
